requirements.txt
Django==3.2.18
django_debug_toolbar==4.1.0
python-decouple==3.8
Jinja2==3.1.2
django-autoslug==1.9.9
django-modeltranslation==0.18.9
psycopg2-binary==2.9.6
phonenumbers==8.13.9
django-phonenumber-field==7.0.2
cryptography==40.0.2
defusedxml==0.7.1
django-allauth==0.54.0
oauthlib==3.2.2
pyjwt==2.7.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
django-imagekit==4.1.0
Pillow==9.5.0
django-filter==23.2
django-user-agents==0.4.0
django-admin-inline-paginator==0.4.0
django-smart-selects==1.6.0
celery==5.2.7
django-celery-results==2.5.1
django-celery-beat==2.5.0
django-money[exchange]==3.1.0
requests==2.31.0
django-cleanup==7.0.0